refactor write barrier and related sanity checks
